LayoutTests/platform/ios-simulator/editing/shadow/execcommand-indent-in-shadow-expected.txt

CONSOLE MESSAGE: line 3: HierarchyRequestError: DOM Exception 3: A Node was inserted somewhere it doesn't belong.
In this test, we do execCommand('Indent') in the direct child of ShadowRoot to confirm a crash doesn't happen.




FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
